added additional tests and documentation
