+++ README.md
@@ -5,7 +5,8 @@ This repository contains the smart contacts for an implementation of the Blend P
 ## Documentation
 
 To learn more about the Blend Protocol, visit the docs:
-* [Blend Docs](https://docs.blend.capital/)
+
+- [Blend Docs](https://docs.blend.capital/)
 
 ## Audits
 
@@ -30,13 +31,21 @@ make test
 The `make` command creates an optimized and un-optimized set of WASM contracts. It's recommended to use the optimized version if deploying to a network.
 
 These can be found at the path:
+
 ```
 target/wasm32-unknown-unknown/optimized
 ```
+
 For help with deployment to a network, please visit the [Blend Utils](https://github.com/blend-capital/blend-utils) repo.
+
+## Contributing
+
+Notes for contributors:
+
+- Under no circumstances should the "overflow-checks" flag be removed otherwise contract math will become unsafe
 
 ## Community Links
 
 A set of links for various things in the community. Please submit a pull request if you would like a link included.
 
-* [Blend Discord](https://discord.com/invite/a6CDBQQcjW)
+- [Blend Discord](https://discord.com/invite/a6CDBQQcjW)
